+++ scripts/ci/osx/before_install.sh
@@ -3,9 +3,10 @@ set -ev
 
 # Prepare build environment
 mkdir dependencies
-mkdir install
-mkdir install/artifacts
 mkdir build
+mkdir build-static
+mkdir install
+mkdir install-static
 
 export VLCQT_VERSION=`cat VERSION`