+++ tests/qml/TestQml.cpp
@@ -18,4 +18,11 @@
 
 #include <QtQuickTest/quicktest.h>
 
+#include "TestsConfig.h"
+
+#if STATIC_BUILD
+    #include <QtPlugin>
+    Q_IMPORT_PLUGIN(VLCQtPlugin)
+#endif
+
 QUICK_TEST_MAIN(player)